REF: Move mongo MDS into headersource.
